requirements.txt
asgiref==3.8.1
Django==5.2
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
